refactor(core-components-amount): renaming
