+++ packages/amount/src/component.tsx
